requirements.txt
aws-cdk-lib==2.84.0
constructs>=10.0.0,<11.0.0
cdk-dynamo-table-view
urllib3>=2.0.3
